Pay Split Manager, charges and pay splits were being shown with zero amounts sometimes. Backported to: 15.2.22.0;15.3.16.0 Reviewed by: Allen - Note: Adding/subtracting doubles cause accuracy errors in the result.  The changes made to the PaySplitEdit window (changing data types from double to decimal) are likely temporary.
git-svn-id: https://opendentalsoft.com:23793/svn/opendental/opendental15.2@12929 098d0599-7f2b-0410-95a5-ad30ac85393a

+++ OpenDental/Forms/FormPayment.cs
@@ -1791,7 +1791,7 @@ namespace OpenDental {
 
 		private void butSplitManage_Click(object sender,EventArgs e) {
 			FormPaySplitManage FormPSM=new FormPaySplitManage();
-			FormPSM.PaymentAmt=PIn.Double(textAmount.Text);
+			FormPSM.PaymentAmt=PIn.Decimal(textAmount.Text);
 			FormPSM.FamCur=Patients.GetFamily(PatCur.PatNum);
 			FormPSM.PatCur=PatCur;
 			FormPSM.PaymentCur=PaymentCur;
@@ -1962,7 +1962,7 @@ namespace OpenDental {
 				//We need to ask if they want to autosplit the payment to start getting procedures associated to splits.
 				if(MsgBox.Show(this,MsgBoxButtons.YesNo,"Would you like to autosplit the payment to outstanding family balances?")) {
 					FormPaySplitManage FormPSM=new FormPaySplitManage();
-					FormPSM.PaymentAmt=PIn.Double(textAmount.Text);
+					FormPSM.PaymentAmt=PIn.Decimal(textAmount.Text);
 					FormPSM.FamCur=Patients.GetFamily(PatCur.PatNum);
 					FormPSM.PatCur=PatCur;
 					FormPSM.PaymentCur=PaymentCur;
@@ -1971,7 +1971,7 @@ namespace OpenDental {
 					FormPSM.ListSplitsCur=new List<PaySplit>();
 					if(FormPSM.ShowDialog()==DialogResult.OK) {
 						SplitList=FormPSM.ListSplitsCur;
-						PaymentCur.PayAmt=FormPSM.AmtTotal;
+						PaymentCur.PayAmt=(double)FormPSM.AmtTotal;
 						if(SplitList.Count==0) {//If they clicked OK without any splits being added, add one split.
 							AddOneSplit();
 						}
